backend/requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-cors-headers==4.0.0
django-phonenumber-field==7.2.0
django-stubs==4.2.6
django-stubs-ext==4.2.5
django-templated-mail==1.1.1
django-types==0.19.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==3.14.4
djoser==2.2.0
docopt==0.6.2
google-auth==2.23.3
google-auth-oauthlib==1.1.0
gspread==5.12.0
idna==3.4
mypy==1.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
phonenumbers==8.13.23
phonenumberslite==8.13.23
pipreqs==0.4.13
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
types-psycopg2==2.9.21.14
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
types-requests==2.31.0.10
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
yarg==0.1.9